Refactor system type api logic #1493
