+++ src/api/systems.test.tsx
@@ -3,7 +3,6 @@ import { MockInstance } from 'vitest';
 import { CopyToSystem, MoveToSystem } from '../app.types';
 import SystemBreadcrumbsJSON from '../mocks/SystemBreadcrumbs.json';
 import SystemsJSON from '../mocks/Systems.json';
-import SystemTypesJSON from '../mocks/SystemTypes.json';
 import { hooksWrapperWithProviders } from '../testUtils';
 import { imsApi } from './api';
 import {
@@ -19,7 +18,6 @@ import {
   useGetSystemIds,
   useGetSystems,
   useGetSystemsBreadcrumbs,
-  useGetSystemTypes,
   useMoveToSystem,
   usePatchSystem,
   usePostSystem,
@@ -76,20 +74,6 @@ describe('System api functions', () => {
           (system) => system.parent_id === '65328f34a40ff5301575a4e3'
         )
       );
-    });
-  });
-
-  describe('useGetSystemTypes', () => {
-    it('sends request to fetch all system types and returns successful response', async () => {
-      const { result } = renderHook(() => useGetSystemTypes(), {
-        wrapper: hooksWrapperWithProviders(),
-      });
-
-      await waitFor(() => {
-        expect(result.current.isSuccess).toBeTruthy();
-      });
-
-      expect(result.current.data).toEqual(SystemTypesJSON);
     });
   });
