+++ src/items/itemsLandingPage.component.tsx
@@ -16,7 +16,8 @@ import { useGetCatalogueCategory } from '../api/catalogueCategories';
 import { useGetCatalogueItem } from '../api/catalogueItems';
 import { useGetItem } from '../api/items';
 import { useGetManufacturer } from '../api/manufacturers';
-import { useGetSystem, useGetSystemType } from '../api/systems';
+import { useGetSystem } from '../api/systems';
+import { useGetSystemType } from '../api/systemTypes';
 import { useAuthorisationState } from '../authProvider.component';
 import ActionMenu from '../common/actionMenu.component';
 import PrimaryImage from '../common/images/primaryImage.component';
