+++ src/systems/systemsTableView.component.tsx
@@ -14,7 +14,7 @@ import {
 import { MRT_Localization_EN } from 'material-react-table/locales/en';
 import React from 'react';
 import { System } from '../api/api.types';
-import { useGetSystemTypes } from '../api/systems';
+import { useGetSystemTypes } from '../api/systemTypes';
 import type { SystemTableType } from '../app.types';
 import {
   COLUMN_FILTER_FUNCTIONS,
